requirements.txt
PyQt6>=6.0
pydantic>=2.0
pydantic-settings>=2.0
pywin32>=300
PyYAML>=6.0
requests>=2.0
pillow>=10.0
win11toast>=0.30
pyinstaller>=6.0
winrt-runtime>=3.0
winrt-windows-devices-bluetooth>=3.0
winrt-Windows.Devices==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
pyinstaller~=6.19.0
winshell>=0.1.1
pyqt-toast-notification>=0.1.0
PyQt6-WebEngine>=6.11.0
markdown~=3.10.2